// shared/Types/IProduct.d.ts
import mongoose from '../../api/node_modules/mongoose'
// Generated by https://quicktype.io

export type IProduct = {
	_id: mongoose.Types.ObjectID
	description: string
	title: string
	image: string
	price: number
	user: mongoose.Types.ObjectID
	colors: string[]
	__v: number
	averageRating: number
	numOfReviews: number
}